refactor(ui): декомпозирован экран списка статей
ArticleListView разделён на отдельные компоненты ArticleListContentView, ArticleListRefreshBanner, ArticleListRowView, ArticleListSectionHeaderView и ArticleListPreviewData
из основного View вынесены layout-части списка, refresh banner, row rendering и preview fixtures, чтобы сократить размер файла и упростить дальнейшую переработку screen-level логики
повторное использование ArticleListItemDTO.updating сохранено в bulk action flow
