+++ RSSReader/Views/ArticleListSectionHeaderView.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+struct ArticleListSectionHeaderView: View {
+    let title: String
+
+    var body: some View {
+        Text(title)
+            .font(.footnote.weight(.semibold))
+            .foregroundStyle(.secondary)
+            .textCase(nil)
+            .frame(maxWidth: .infinity, alignment: .leading)
+    }
+}
